+++ resources/user.rb
@@ -20,6 +20,7 @@ unified_mode true
 
 property :password,
         String,
+        sensitive: true,
         description: 'User password for samba and the system'
 
 property :comment,
@@ -55,7 +56,7 @@ end
 
 action :create do
   user new_resource.name do
-    password generate_system_password
+    password new_resource.password
     comment new_resource.comment
     home new_resource.home
     shell new_resource.shell
@@ -95,14 +96,5 @@ action :delete do
     execute "Delete #{new_resource.name}" do
       command "smbpasswd -x #{new_resource.name}"
     end
-  end
-end
-
-action_class.class_eval do
-  def generate_system_password
-    system_password = \
-      new_resource.password.crypt('$6$' + SecureRandom.random_number(36**8).to_s(36))
-    Chef::Log.debug "SC: generated system password: #{system_password}"
-    system_password
   end
 end