+++ test/integration/server/server_spec.rb
@@ -36,11 +36,10 @@ describe file('/etc/samba/smb.conf') do
 end
 
 describe file('/etc/samba/smb.conf') do
-  its('content') { should match(/realm = $/) }
   its('content') { should match(/password server = \*$/) }
   its('content') { should match(/kerberos method = secrets only$/) }
   its('content') { should match(/workgroup = HOME/) }
-  its('content') { should match(/server string = fat/) }
+  its('content') { should match(/server string = foxtrot/) }
   its('content') { should match(/security = user/) }
   its('content') { should match(/map to guest = Bad User/) }
   its('content') { should match(/interfaces = lo 127.0.0.1 enp\* eth\*/) }