+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Samba Changelog
 
-## Unreleased
+## 2.0.0 - *2021-05-11*
 
 - Remove unused properties shares & enable_users_search
 - Remove recipes